requirements.txt
dash[diskcache]==2.17.1
dash_auth==2.3.0
dash-bootstrap-components==1.6.0
pandas==2.2.2
pydantic==2.8.2
werkzeug==3.0.3
snakeviz==2.2.0
gunicorn==22.0.0
pyyaml==6.0.1
